requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
dj-database-url==2.3.0
Django==5.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.32.0
whitenoise==6.7.0
